Detect expired tokens on post requests
Many post requests return an empty body so we don't parse them as JSON.
A significant exception is when a post is made with an expired token.
In this case the response is a JSON body that indicates the token is
expired.  In this case we need to manually handle the JSON.parse since
we can not tell the request library to do it.